backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2
google-genai